refactor: group compaction logic into sessionCompactor collaborator
Move compactWithReason, compactIfNeeded, preCompactSourceFor, and
joinPrompts into a sessionCompactor struct. The LocalRuntime holds a
*sessionCompactor and delegates all compaction calls to it. This
concentrates compaction reasoning (previously spread across runtime.go,
loop.go, loop_steps.go, and session_compaction.go) in one place.
